CDRIVER-2846 OpenSSL 1.1.1 compatibility
OpenSSL 1.1.1 supports TLSv1.3, and the docs say "TLSv1.3 sends more
non-application data records after the handshake is finished." When the async
loop detects a socket is readable, the available data might be non-application
data and BIO_read returns 0. Update the async code to keep trying to read in
this scenario.

Also fix the CMake config check for ASN1_STRING_get0_data, and avoid
deprecation warnings with OpenSSL 1.1.0+ from the old ASN1_STRING_data.

Finally, if we test build an old OpenSSL locally on a system where 1.1.1 is the
default, we must keep the old OpenSSL out of CMake's own LD path or it will
fail to start.